@@ -11,38 +11,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
+    "@angular/animations": "^8.2.6",
+    "@angular/cdk": "^8.2.0",
+    "@angular/common": "~8.2.6",
+    "@angular/compiler": "~8.2.6",
+    "@angular/core": "^8.2.6",
+    "@angular/forms": "~8.2.6",
+    "@angular/material": "^8.2.0",
+    "@angular/platform-browser": "~8.2.6",
+    "@angular/platform-browser-dynamic": "~8.2.6",
+    "@angular/router": "~8.2.6",
+    "angular-material-sass-files": "^1.0.0",
+    "angular2-datetimepicker": "^1.1.1",
+    "core-js": "^3.2.1",
+    "rxjs": "~6.5.3",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^0.13.0",
+    "@angular/compiler-cli": "~8.2.6",
+    "@angular/language-service": "~8.2.6",
+    "@types/node": "~12.7.5",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.20.0",
+    "typescript": "~3.4.0"
   }
 }
